Added joystick control for ball commands

+++ src/pico-ball-controller/src/mbot.h
@@ -27,7 +27,7 @@
 
 // need to be packed in order to make the byte copy from numpy and C to play nicely
 typedef struct __attribute__((__packed__ )) mbot_motor_command{
-    uint8_t intake; // 1 is on, 0 is off
+    float intake; // 1 is on, 0 is off
     uint8_t shoot_ball; // 1 is shoot, 0 is don't; If 1, this overrides intake
 } mbot_ball_command_t;
